insta file

+++ controllers/imageController.js
@@ -4,7 +4,7 @@ import Comment from "../models/Comment";
 
 export const home = async (req, res) => {
   try {
-    const images = await Image.find({}).sort({ _id: -1 });
+    const images = await Image.find({}).sort({ _id: -1 }).populate("creator");
     res.render("home", { pageTitle: "Home", images });
   } catch (error) {
     console.log(error);
@@ -18,14 +18,16 @@ export const getUpload = (req, res) =>
 
 export const postUpload = async (req, res) => {
   const {
-    body: { description, createdAt },
+    body: { description },
     file: { path },
   } = req;
   const newImage = await Image.create({
     fileUrl: path.replace(/\\/g, "/"),
     description,
-    createdAt,
+    creator: req.user.id,
   });
+  req.user.images.push(newImage.id);
+  req.user.save();
   res.redirect(routes.imageDetail(newImage.id));
 };
 
@@ -35,11 +37,15 @@ export const imageDetail = async (req, res) => {
     params: { id },
   } = req;
   try {
-    const image = await Image.findById(id); //위에서 params로 얻은 id
-    //.populate("creator")
-    //.populate("comments");
+    const image = await Image.findById(id) //위에서 params로 얻은 id
+      .populate("creator")
+      .populate("comments");
+    image.views += 1; //view 수 증가
+    image.save();
+    res.status(200);
     res.render("imageDetail", { pageTitle: "ImageDetail", image });
   } catch (error) {
+    res.status(400);
     res.redirect(routes.home);
   }
 };
@@ -52,7 +58,11 @@ export const getEditImage = async (req, res) => {
 
   try {
     const image = await Image.findById(id);
-    res.render("editImage", { pageTitle: "edit Image", image });
+    if (String(image.creator) !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editImage", { pageTitle: "edit Image", image });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -77,12 +87,56 @@ export const deleteImage = async (req, res) => {
     params: { id },
   } = req;
   try {
-    await Image.findOneAndRemove({ _id: id });
+    const image = await Image.findById(id);
+    if (String(image.creator) !== req.user.id) {
+      throw Error();
+    } else {
+      await Image.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
   res.redirect(routes.home);
 };
 
-//댓글 삭제
+// views 기능
+export const postRegisterLike = async (req, res) => {
+  const {
+    params: { id },
+  } = req;
+  try {
+    const image = await Image.findById(id);
+    image.views += 1;
+    image.save();
+    res.status(200);
+  } catch (error) {
+    res.status(400);
+  } finally {
+    res.end();
+  }
+};
+
+// 댓글 달기
+export const postAddComment = async (req, res) => {
+  const {
+    params: { id },
+    body: { comment },
+    user,
+  } = req;
+  try {
+    const image = await Image.findById(id);
+    const newComment = await Comment.create({
+      text: comment,
+      creator: user.id,
+    });
+    image.comments.push(newComment._id);
+    image.save();
+  } catch (error) {
+    res.status(400);
+  } finally {
+    res.end();
+  }
+};
+
+// 댓글 삭제
 export const deleteComment = (req, res) => res.render("deleteComment");
